backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-multipart
redis
httpx
pytest
